backend: exception handling

+++ backend/src/main/java/com/users/location/backend/User/UserService.java
@@ -1,5 +1,7 @@
 package com.users.location.backend.User;
 
+import com.users.location.backend.Exceptions.BadRequest;
+import com.users.location.backend.Exceptions.NotFound;
 import com.users.location.backend.User.DTO.UserDTO;
 import com.users.location.backend.User.Entities.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,23 +19,32 @@ public class UserService {
     return userRepository.findAll();
   }
 
-  public Optional<User> getUserById(Long id) {
-    return userRepository.findById(id);
+  public User getUserById(Long id) {
+    return userRepository.findById(id).orElseThrow(() -> new NotFound("User not found"));
   }
 
   public User createUser(UserDTO userDTO) {
+    if (userRepository.findByEmail(userDTO.getEmail()).isPresent()) {
+      throw new BadRequest("A user with the provided email already exists.");
+    }
+
     User user = User.builder()
-            .name(userDTO.getName())
-            .email(userDTO.getEmail())
-            .address(userDTO.getAddress())
-            .latitude(userDTO.getLatitude())
-            .longitude(userDTO.getLongitude())
-            .build();
+        .name(userDTO.getName())
+        .email(userDTO.getEmail())
+        .address(userDTO.getAddress())
+        .latitude(userDTO.getLatitude())
+        .longitude(userDTO.getLongitude())
+        .build();
     return userRepository.save(user);
   }
 
   public User updateUser(Long id, UserDTO userDTO) {
-    User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+    User user = userRepository.findById(id).orElseThrow(() -> new NotFound("User not found"));
+    Optional<User> existingUserWithEmail = userRepository.findByEmail(userDTO.getEmail());
+    if (existingUserWithEmail.isPresent() && !existingUserWithEmail.get().getId().equals(id)) {
+      throw new BadRequest("Email is already taken by another user.");
+    }
+
     user.setName(userDTO.getName());
     user.setEmail(userDTO.getEmail());
     user.setAddress(userDTO.getAddress());
@@ -43,6 +54,10 @@ public class UserService {
   }
 
   public void deleteUser(Long id) {
+    if (!userRepository.existsById(id)) {
+      throw new NotFound("User not found");
+    }
+
     userRepository.deleteById(id);
   }
 }
